Move non-business types to kubernetes layer
